adding cobra and cobra-ifying the CLI
