+++ cmd/cli/init.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"log"
+)
+
+func init() {
+	log.SetFlags(log.Flags() | log.Lshortfile)
+}
